third_party/blink/web_tests/fast/events/touch/gesture/gesture-tap-result-expected.txt


This is plain text with no handler
This text consumes events using preventDefault()

Clicking or tapping on the "consumes" section should have no effect on the selection, but clicking in the plain section should clear it.

This tests Tap events being consumed by a handler.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

PASS consumesResult is true
PASS window.getSelection().toString() is not 
PASS plainResult is false
PASS window.getSelection().toString() is ""
